Consolidate Native AOT integration coverage

Copilot-Session: b24622c7-c9d6-44f0-9a20-85aac5ff11d6
