+++ .github/workflows/build.yml
@@ -311,7 +311,7 @@ jobs:
       run: dotnet publish -c Release -r win-x64
       working-directory: ${{ runner.temp }}/integration-tests/buildtask-sdk
     - name: 🧪 Run Native AOT automatic WinRT marshalling
-      run: ./auto-winrt-aot/bin/Release/net9.0-windows10.0.22621.0/win-x64/publish/buildtask-sdk-auto-winrt-aot.exe --validate-com-out-ptr
+      run: ./app/bin/Release/net9.0-windows10.0.22621.0/win-x64/publish/buildtask-sdk.exe --validate-com-out-ptr
       working-directory: ${{ runner.temp }}/integration-tests/buildtask-sdk
       shell: pwsh
