+++ azure-pipelines/integration-test.yml
@@ -62,7 +62,7 @@ jobs:
     displayName: 🧪 dotnet publish
     workingDirectory: $(Pipeline.Workspace)\integration-tests\buildtask-sdk
     condition: eq(variables.currentSdk, 'true')
-  - pwsh: .\auto-winrt-aot\bin\Release\net9.0-windows10.0.22621.0\win-x64\publish\buildtask-sdk-auto-winrt-aot.exe --validate-com-out-ptr
+  - pwsh: .\app\bin\Release\net9.0-windows10.0.22621.0\win-x64\publish\buildtask-sdk.exe --validate-com-out-ptr
     displayName: 🧪 Run Native AOT automatic WinRT marshalling
     workingDirectory: $(Pipeline.Workspace)\integration-tests\buildtask-sdk
     condition: eq(variables.currentSdk, 'true')
