+++ integration-tests/buildtask-sdk/app/NativeMethods.json
@@ -1,5 +1,5 @@
 {
-  "$schema": "..\\..\\..\\src\\Microsoft.Windows.CsWin32\\settings.schema.json",
+  "$schema": "https://aka.ms/CsWin32.schema.json",
   "allowMarshaling": true,
   "public": false,
   "comInterop": {
