+++ integration-tests/buildtask-sdk/app/NativeMethods.txt
@@ -1,1 +1,3 @@
 ﻿GetTickCount
+IShellItem
+SHCreateItemFromParsingName
