+++ integration-tests/buildtask-sdk/app/buildtask-sdk.csproj
@@ -2,8 +2,9 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net9.0-windows</TargetFramework>
-    <LangVersion>9</LangVersion>
+    <TargetFramework>net9.0-windows10.0.22621.0</TargetFramework>
+    <!-- Generated COM interfaces require C# 12. The dependency project retains C# 9 coverage. -->
+    <LangVersion>12</LangVersion>
     <PublishAot>true</PublishAot>
     <CsWin32RunAsBuildTask>true</CsWin32RunAsBuildTask>
     <DisableRuntimeMarshalling>true</DisableRuntimeMarshalling>
