--- integration-tests/buildtask-sdk/auto-winrt-aot/NativeMethods.txt
@@ -1,4 +1,0 @@
-CreateBindCtx
-IStream
-IShellItem
-SHCreateItemFromParsingName
